requirements.txt
tda-api==1.3.5
bokeh==2.3.2
